% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abortSubmission.R
\name{abortSubmission}
\alias{abortSubmission}
\title{Abort Terra Submission}
\usage{
abortSubmission(accountEmail, billingProjectName, workspaceName, submissionId)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission ID}
}
\description{
Abort Terra Submission
}
