% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobakery_currentInput.R
\name{biobakery_currentInput}
\alias{biobakery_currentInput}
\title{Check the current input arguments}
\usage{
biobakery_currentInput(
  accountEmail,
  billingProjectName,
  workspaceName,
  allConfig = FALSE
)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{allConfig}{Under the default (\code{FALSE}), the file path to the input
files list and the paths to input files will be returned. If it's set to
\code{TRUE}, the whole method configuration will be returned.}
}
\description{
Check the current input arguments
}
