% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneWorkspace.R
\name{cloneWorkspace}
\alias{cloneWorkspace}
\title{Clone bioBakery workflow template}
\usage{
cloneWorkspace(accountEmail, projectName, workspaceName)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{projectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}
}
\description{
Clone bioBakery workflow template
}
\examples{
cloneWorkspace(accountEmail = "shbrief@gmail.com",
              projectName = "waldronlab-terra-rstudio",
              workspaceName = "mtx_workflow_biobakery_ver3")

}
