% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentInput.R
\name{currentInput}
\alias{currentInput}
\title{Check the current input arguments}
\usage{
currentInput(accountEmail, billingProject, workspaceName, allInputs = FALSE)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProject}{Name of the billing project}

\item{workspaceName}{Name of the workspace}
}
\description{
Check the current input arguments
}
