% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{Download output}
\usage{
getOutput(
  accountEmail,
  billingProjectName,
  workspaceName,
  submissionId = NULL,
  keyword = NULL,
  dest_dir = "."
)
}
\description{
Download output
}
\examples{
fname <- "wmgx_report.pdf"


}
