% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchWorkflow.R
\name{launchWorkflow}
\alias{launchWorkflow}
\title{Launch workflow on Terra}
\usage{
launchWorkflow(accountEmail, billingProject, workspaceName)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProject}{Name of the billing project}

\item{workspaceName}{Name of the workspace}
}
\description{
Launch workflow on Terra
}
