% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOutput.R
\name{listOutput}
\alias{listOutput}
\title{List outputs}
\usage{
listOutput(
  accountEmail,
  projectName,
  workspaceName,
  submissionId = NULL,
  keyword = NULL
)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{projectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission Id. If it's not provided, the most recent
submission id will be used.}

\item{keyword}{A keyword in the output file name. Under the default \code{NULL},
names of all the output files with \code{.tsv} extension will be returned.}
}
\description{
List outputs
}
