% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorSubmission.R
\name{monitorSubmission}
\alias{monitorSubmission}
\title{Check the status of submitted jobs}
\usage{
monitorSubmission(
  accountEmail,
  billingProjectName,
  workspaceName,
  mostRecentOnly = TRUE
)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{mostRecentOnly}{Under the default (\code{TRUE}), the status of only the
lastest submission is returned. If it's set to \code{FALSE}, all the submission
status information will be returned in a data frame.}
}
\value{
A list or a data frame containing the submission information, such as
submission status, submission date, and submission Id.
}
\description{
Check the status of submitted jobs
}
\examples{
monitorSubmission(accountEmail = "shbrief@gmail.com",
                  billingProjectName = "waldronlab-terra-rstudio",
                  workspaceName = "mtx_workflow_biobakery_ver3")

}
