% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobakery_currentInput.R
\name{biobakery_currentInput}
\alias{biobakery_currentInput}
\title{Check the current input arguments}
\usage{
biobakery_currentInput(
  workspaceName,
  allConfig = FALSE,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)
}
\arguments{
\item{workspaceName}{Name of the workspace}

\item{allConfig}{Under the default (\code{FALSE}), the file path to the
input files list and the paths to input files will be returned. If it's
set to \code{TRUE}, the whole method configuration will be returned.}

\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}
}
\description{
Check the current input arguments
}
