% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchWorkflow.R
\name{launchWorkflow}
\alias{launchWorkflow}
\title{Launch workflow on Terra}
\usage{
launchWorkflow(accountEmail, projectName, workspaceName)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{projectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}
}
\description{
Launch workflow on Terra
}
\examples{
launchWorkflow(accountEmail = "shbrief@gmail.com",
               projectName = "waldronlab-terra-rstudio",
               workspaceName = "mtx_workflow_biobakery_ver3")

}
