% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOutput.R
\name{listOutput}
\alias{listOutput}
\title{List outputs}
\usage{
listOutput(
  accountEmail,
  billingProjectName,
  workspaceName,
  submissionId = NULL,
  keyword = NULL,
  all = FALSE
)
}
\arguments{
\item{accountEmail}{Email linked to Terra account}

\item{billingProjectName}{Name of the billing project}

\item{workspaceName}{Name of the workspace}

\item{submissionId}{Submission Id. If it's not provided, the most recent
submission id will be used.}

\item{keyword}{A character string containing a regular exrpession to be matched
in the output file name. Under the default \code{NULL}, the names of all the output
files with \code{.tsv} extension will be returned.}

\item{all}{A logical. If it's set to \code{TRUE}, \code{keyword} argument will
be ignored and all the outputs from the workflow, including log files, will be
returned. Default is (\code{FALSE}).}
}
\value{
A tibble with four columns
\itemize{
    \item file : character() 'base name' of the file in the bucket.
    \item workflow : character() name of the workflow the file is associated with.
    \item task : character() name of the task in the workflow that generated the file.
    \item path : charcter() full path to the file in the google bucket.
}
}
\description{
List outputs
}
